requirements.txt
pyserial==3.5
pynmea2==1.19.0
requests==2.32.3

